refactoring prx tests
